fix: sales role detection hardening and domain-aware scoring ceiling (#582)
* feat: strict evidence-based resume analysis scoring

Backend: add industry-db score cap, related-exp weight, recommendation
thresholds, numeric breakdown parsing, and non-employer brand/company
hit detection. API: refactor clear-analyses into paginated batch helper.
Sales role: filter generic company boilerplate matches.
Frontend: add recommendationFromScore(), update scoring pipeline to
bump/bypass industry-db V2 based on strict evidence flags.
Tests: add tests for analyze.ts, ingest-compute-service, scoring,
SearchHero, useIndustryKeywords, useResumeListState, resumes route,
ResumeSearchPage.

* update

* update

+++ packages/cli/internal/client/api.go
@@ -623,7 +623,7 @@ type AnalysisTaskConfig struct {
 type AnalysisTask struct {
 	ID           string               `json:"_id"`
 	Status       string               `json:"status"`
-	CreatedAt    int64                `json:"_creationTime"`
+	CreatedAt    float64              `json:"_creationTime"`
 	Config       *AnalysisTaskConfig  `json:"config,omitempty"`
 	Progress     *AnalysisTaskProgress `json:"progress,omitempty"`
 	Results      *AnalysisTaskResults  `json:"results,omitempty"`
